pyproject.toml
[tool.poetry]
name = "drf-error-issue"
version = "0.1.0"
description = ""
authors = ["Brent O'Connor <epicserve@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
django-filter = "^2.3.0"
pytest-django = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
